+++ src/index.ts
@@ -3,7 +3,7 @@ export * from './geometries/geometryCollection';
 export * from './geometries/line';
 export * from './geometries/point';
 export * from './geometries/polygon';
-export * from './measurements/measurements';
+export * from './measurements';
 export * from './tiles/tile';
 export * from './tiles/tileMatrixSet';
